+++ packages/suspense-website/src/examples/isPromiseLike/util.ts
@@ -5,17 +5,17 @@ const commentsCache = {
 const userId = "123";
 
 // REMOVE_BEFORE
-import { isThenable } from "suspense";
+import { isPromiseLike } from "suspense";
 
 function UserComments({ userId }: { userId: string }) {
   let comments: string[];
   try {
     // Try to load comments using Suspense
     comments = commentsCache.fetchUserCommentsSuspense(userId);
-  } catch (errorOrThenable) {
-    if (isThenable(errorOrThenable)) {
+  } catch (errorOrPromiseLike) {
+    if (isPromiseLike(errorOrPromiseLike)) {
       // Re-throw thenables so that Suspense works
-      throw errorOrThenable;
+      throw errorOrPromiseLike;
     } else {
       // Ignore other types of errors though
       comments = [];