+++ packages/suspense-website/src/routes/api/isPromiseLike.tsx
@@ -1,7 +1,7 @@
 import Container from "../../components/Container";
 import Code from "../../components/Code";
 import Header from "../../components/Header";
-import { isThenable } from "../../examples";
+import { isPromiseLike } from "../../examples";
 import Block from "../../components/Block";
 import Note from "../../components/Note";
 import { ExternalLink } from "../../components/ExternalLink";
@@ -11,16 +11,17 @@ export default function Route() {
   return (
     <Container>
       <Block>
-        <Header title="isThenable" />
+        <Header title="isPromiseLike" />
       </Block>
       <Note type="warn">
-        Although this package exports the <code>isThenable</code> API, you
+        Although this package exports the <code>isPromiseLike</code> API, you
         probably won't need to use it directly in most cases.
       </Note>
       <Block>
         <p>
-          "Thenable" are a lower level concept that the caches in this package
-          are built on top of. To quote{" "}
+          Things that are "promise like" (also sometimes called "thenables") are
+          a lower level concept that the caches in this package are built on top
+          of. To quote{" "}
           <ExternalLink to="https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise#thenables">
             the MDM docs
           </ExternalLink>
@@ -28,13 +29,13 @@ export default function Route() {
         </p>
         <Note title="Thenables" type="quote">
           <p>
-            A thenable implements the <code>.then()</code> method, which is
+            A "thenable" implements the <code>.then()</code> method, which is
             called with two callbacks: one for when the promise is fulfilled,
             one for when it's rejected. Promises are thenables as well.
           </p>
         </Note>
         <p>
-          <code>isThenable</code> can be used to determine if a value is a
+          <code>isPromiseLike</code> can be used to determine if a value is a
           "thenable" or not.
         </p>
       </Block>
@@ -44,7 +45,7 @@ export default function Route() {
           You can use this utility method to ignore failures for a particular
           Suspense cache without also swallowing other types of errors.
         </p>
-        <Code code={isThenable.util} />
+        <Code code={isPromiseLike.util} />
       </Block>
     </Container>
   );