+++ packages/suspense/src/index.ts
@@ -8,7 +8,7 @@ export * from "./hooks/useCacheStatus";
 export * from "./hooks/useStreamingValues";
 export * from "./utils/createDeferred";
 export * from "./utils/createInfallibleCache";
-export * from "./utils/isThenable";
+export * from "./utils/isPromiseLike";
 export * from "./utils/parallelize";
 
 // TypeScript types