+++ packages/suspense/src/utils/isThenable.ts
@@ -1,5 +1,0 @@
-import { Thenable } from "../types";
-
-export function isThenable(value: any): value is Thenable<any> {
-  return value != null && typeof value.then === "function";
-}